backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
qrcode==7.4.2
reportlab==4.2.2
orjson==3.10.7
